Add files via upload
This is the first test commit. Here, I have imported the libraries and created a Class called Node. I have also created the cost comparing parameter that I am planning to use in the implementation of Dijkstra Algorithm. I have also created basic actions of the node as different functions and assigned appropriate costs to them.