Map Update, Dijkstra and Plotter Implementation
I faced issues with the location and buffer region of the hexagon and I tried to write a rough implementation of Dijkstra Algorithm. I created a new class for Plotting path and Obstacle space for visualization. The issues with this code is that the code works perfectly fine when given smaller inputs but does not compute end of corner inputs. The ram usage of my computer reaches 80% without any plot.